requirements.txt
# ============================================================
# ScribbleNet - Dependencies
# ============================================================

# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Frontend
streamlit>=1.28.0

# Visualization
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-Levenshtein>=0.21.0

# PDF generation
fpdf2>=2.7.0

# Logging
colorlog>=6.7.0
